health: #184 moved types Expand & UnfoldedExpansion to common lib
