+++ packages/common/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./lib/entity.type";
+export * from "./lib/expand.type";
 export * from "./lib/expansion-value.type";
